+++ sql/mysql/create.tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `clients` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT(20) NOT NULL AUTO_INCREMENT,
   `lastName` varchar(255) NOT NULL,
   `firstName` varchar(255) NOT NULL,
   `dateOfBirth` date NOT NULL,
@@ -7,8 +7,8 @@ CREATE TABLE `clients` (
 ) ENGINE=InnoDB AUTO_INCREMENT=5 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 
 CREATE TABLE `accounts` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `clientId` int(11) NOT NULL,
+  `id` BIGINT(20) NOT NULL AUTO_INCREMENT,
+  `clientId` BIGINT(20) NOT NULL,
   `number` int(7) NOT NULL,
   `type` tinyint(2) NOT NULL,
   `amount` decimal(10, 2) NOT NULL,
@@ -20,9 +20,9 @@ CREATE TABLE `accounts` (
 ) ENGINE=InnoDB AUTO_INCREMENT=2 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 
 CREATE TABLE `cards` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `accountId` int(11) NOT NULL,
-  `card_number` int(16) NOT NULL,
+  `id` BIGINT(20) NOT NULL AUTO_INCREMENT,
+  `accountId` BIGINT(20) NOT NULL,
+  `card_number` BIGINT(16) NOT NULL,
   `owner_name` varchar(510) NOT NULL,
   `cvv` smallint(4) NOT NULL,
   `exp_year` year(4) NOT NULL,
